+++ src/main/java/nextstep/subway/line/advice/LineControllerAdvice.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import nextstep.subway.common.exception.ResourceNotFoundException;
-import nextstep.subway.line.exception.ResourceAlreadyExistException;
+import nextstep.subway.common.exception.ResourceAlreadyExistException;
 import nextstep.subway.line.ui.LineController;
 
 @RestControllerAdvice(assignableTypes = LineController.class)
